requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
playwright
beautifulsoup4
lxml
sentence-transformers
faiss-cpu
numpy
google-generativeai